# api/swagger.yaml
openapi: 3.0.0
info:
  title: Portfolio API
  description: Backend API for Tomiwa Aluko's Portfolio
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
paths:
  /health:
    get:
      summary: Health Check
      description: Returns the status of the API server.
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /contact:
    post:
      summary: Send contact message
      description: Sends email via Resend (portfolio contact form).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - message
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                subject:
                  type: string
                message:
                  type: string
      responses:
        '201':
          description: Message sent
        '400':
          description: Validation error
        '503':
          description: Email not configured
  /guestbook:
    get:
      summary: Get Guestbook Entries
      description: Retrieve all approved guestbook entries.
      responses:
        '200':
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    name:
                      type: string
                    message:
                      type: string
                    createdAt:
                      type: string
    post:
      summary: Sign Guestbook
      description: Submit a new entry to the guestbook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - message
              properties:
                name:
                  type: string
                message:
                  type: string
      responses:
        '201':
          description: Entry created successfully